Move utility method from evaluation utilities class to core utilities class.
